Nevermind, Added more autofills
I made it so therefore when the user inputs their card information theres alot of validation that occurs. If the user types something wrong the browser will not allow the use to submit the information because its wrong and a pop up will tell the user that its wrong. If all correct the user may submit his information